♻️ refactor : Admin 메서드 RESTful하게 변경
