+++ src/main/java/com/example/ingle/domain/stamp/controller/StampController.java
@@ -1,8 +1,8 @@
 package com.example.ingle.domain.stamp.controller;
 
 import com.example.ingle.domain.stamp.StampService;
-import com.example.ingle.domain.stamp.res.CompleteTutorialResponse;
-import com.example.ingle.domain.stamp.res.StampResponse;
+import com.example.ingle.domain.stamp.dto.res.CompleteTutorialResponse;
+import com.example.ingle.domain.stamp.dto.res.StampResponse;
 import com.example.ingle.domain.member.domain.MemberDetail;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
